Task: https://jira.onyxenterprises.com/browse/WT-165 Description: add SkillTestType, add qaAgentTechnicalskillsType, implement technicalSkills() method in QaController, implement buildJiraHours() method in qaTechnicalSkillsDTOBuilder, improve technicalSkills view to show data

+++ src/Service/User/QaAgent/QaSkill/QaJiraHoursRow.php
@@ -32,7 +32,7 @@ class QaJiraHoursRow
         $this->requiredHours = $hours;
         return $this;
     }
-    public function getRequiredHours(): int
+    public function getRequiredHours(): float
     {
         return $this->requiredHours;
     }
